repackage http module
