+++ modules/http/src/main/java/org/mule/compatibility/module/http/internal/HttpParamType.java
@@ -1,0 +1,12 @@
+/*
+ * Copyright (c) MuleSoft, Inc.  All rights reserved.  http://www.mulesoft.com
+ * The software in this package is published under the terms of the CPAL v1.0
+ * license, a copy of which has been included with this distribution in the
+ * LICENSE.txt file.
+ */
+package org.mule.compatibility.module.http.internal;
+
+public enum HttpParamType {
+  QUERY_PARAM, URI_PARAM, HEADER;
+
+}
