+++ modules/http/src/main/java/org/mule/compatibility/module/http/internal/request/grizzly/GrizzlyHttpClient.java
@@ -1,0 +1,420 @@
+/*
+ * Copyright (c) MuleSoft, Inc.  All rights reserved.  http://www.mulesoft.com
+ * The software in this package is published under the terms of the CPAL v1.0
+ * license, a copy of which has been included with this distribution in the
+ * LICENSE.txt file.
+ */
+package org.mule.compatibility.module.http.internal.request.grizzly;
+
+import static com.ning.http.client.Realm.AuthScheme.NTLM;
+import static java.lang.Integer.MAX_VALUE;
+import static java.lang.Runtime.getRuntime;
+import static java.util.Objects.requireNonNull;
+import static java.util.concurrent.TimeUnit.SECONDS;
+import static org.mule.runtime.api.i18n.I18nMessageFactory.createStaticMessage;
+import static org.mule.runtime.core.api.scheduler.SchedulerConfig.config;
+import static org.mule.service.http.api.HttpHeaders.Names.CONNECTION;
+import static org.mule.service.http.api.HttpHeaders.Values.CLOSE;
+
+import org.mule.compatibility.module.http.internal.request.HttpClient;
+import org.mule.compatibility.module.http.internal.request.HttpClientConfiguration;
+import org.mule.compatibility.module.http.internal.request.NtlmProxyConfig;
+import org.mule.runtime.api.exception.MuleException;
+import org.mule.runtime.api.lifecycle.InitialisationException;
+import org.mule.runtime.api.scheduler.Scheduler;
+import org.mule.runtime.api.tls.TlsContextFactory;
+import org.mule.runtime.api.tls.TlsContextTrustStoreConfiguration;
+import org.mule.runtime.core.api.DefaultMuleException;
+import org.mule.runtime.core.api.lifecycle.LifecycleUtils;
+import org.mule.runtime.core.api.scheduler.SchedulerService;
+import org.mule.runtime.core.exception.MessagingException;
+import org.mule.runtime.core.execution.CompletionHandler;
+import org.mule.runtime.core.util.IOUtils;
+import org.mule.runtime.core.util.StringUtils;
+import org.mule.service.http.api.client.HttpAuthenticationType;
+import org.mule.service.http.api.client.HttpRequestAuthentication;
+import org.mule.service.http.api.client.proxy.ProxyConfig;
+import org.mule.service.http.api.domain.entity.ByteArrayHttpEntity;
+import org.mule.service.http.api.domain.entity.InputStreamHttpEntity;
+import org.mule.service.http.api.domain.entity.multipart.HttpPart;
+import org.mule.service.http.api.domain.entity.multipart.MultipartHttpEntity;
+import org.mule.service.http.api.domain.message.request.HttpRequest;
+import org.mule.service.http.api.domain.message.response.HttpResponse;
+import org.mule.service.http.api.domain.message.response.HttpResponseBuilder;
+import org.mule.service.http.api.tcp.TcpClientSocketProperties;
+
+import com.ning.http.client.AsyncCompletionHandler;
+import com.ning.http.client.AsyncHttpClient;
+import com.ning.http.client.AsyncHttpClientConfig;
+import com.ning.http.client.ListenableFuture;
+import com.ning.http.client.ProxyServer;
+import com.ning.http.client.Realm;
+import com.ning.http.client.Request;
+import com.ning.http.client.RequestBuilder;
+import com.ning.http.client.Response;
+import com.ning.http.client.generators.InputStreamBodyGenerator;
+import com.ning.http.client.multipart.ByteArrayPart;
+import com.ning.http.client.providers.grizzly.GrizzlyAsyncHttpProvider;
+import com.ning.http.client.providers.grizzly.GrizzlyAsyncHttpProviderConfig;
+
+import java.io.IOException;
+import java.net.InetAddress;
+import java.net.UnknownHostException;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeoutException;
+
+import javax.net.ssl.SSLContext;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class GrizzlyHttpClient implements HttpClient {
+
+  private static final int MAX_CONNECTION_LIFETIME = 30 * 60 * 1000;
+
+  private static final Logger logger = LoggerFactory.getLogger(GrizzlyHttpClient.class);
+
+  private final TlsContextFactory tlsContextFactory;
+  private final ProxyConfig proxyConfig;
+  private final TcpClientSocketProperties clientSocketProperties;
+
+  private int maxConnections;
+  private boolean usePersistentConnections;
+  private int connectionIdleTimeout;
+  private String threadNamePrefix;
+  private Scheduler selectorScheduler;
+  private Scheduler workerScheduler;
+  private SchedulerService schedulerService;
+  private String ownerName;
+
+  private AsyncHttpClient asyncHttpClient;
+  private SSLContext sslContext;
+
+  public GrizzlyHttpClient(HttpClientConfiguration config, SchedulerService schedulerService) {
+    requireNonNull(schedulerService);
+
+    this.tlsContextFactory = config.getTlsContextFactory();
+    this.proxyConfig = config.getProxyConfig();
+    this.clientSocketProperties = config.getClientSocketProperties();
+    this.maxConnections = config.getMaxConnections();
+    this.usePersistentConnections = config.isUsePersistentConnections();
+    this.connectionIdleTimeout = config.getConnectionIdleTimeout();
+    this.threadNamePrefix = config.getThreadNamePrefix();
+    this.ownerName = config.getOwnerName();
+
+    this.schedulerService = schedulerService;
+  }
+
+  @Override
+  public void start() throws MuleException {
+    selectorScheduler = schedulerService
+        .customScheduler(config().withMaxConcurrentTasks(getRuntime().availableProcessors() + 1).withName(threadNamePrefix),
+                         MAX_VALUE);
+    workerScheduler = schedulerService.ioScheduler();
+
+    AsyncHttpClientConfig.Builder builder = new AsyncHttpClientConfig.Builder();
+    builder.setAllowPoolingConnections(true);
+
+    configureTransport(builder);
+
+    configureTlsContext(builder);
+
+    configureProxy(builder);
+
+    configureConnections(builder);
+
+    AsyncHttpClientConfig config = builder.build();
+
+    asyncHttpClient = new AsyncHttpClient(new GrizzlyAsyncHttpProvider(config), config);
+  }
+
+  private void configureTlsContext(AsyncHttpClientConfig.Builder builder) throws MuleException {
+    if (tlsContextFactory != null) {
+      LifecycleUtils.initialiseIfNeeded(tlsContextFactory);
+      try {
+        sslContext = tlsContextFactory.createSslContext();
+      } catch (Exception e) {
+        throw new DefaultMuleException(createStaticMessage("Cannot initialize SSL context"), e);
+      }
+
+      // This sets all the TLS configuration needed, except for the enabled protocols and cipher suites.
+      builder.setSSLContext(sslContext);
+      // These complete the set up
+      if (tlsContextFactory.getEnabledCipherSuites() != null) {
+        builder.setEnabledCipherSuites(tlsContextFactory.getEnabledCipherSuites());
+      }
+      if (tlsContextFactory.getEnabledProtocols() != null) {
+        builder.setEnabledProtocols(tlsContextFactory.getEnabledProtocols());
+      }
+      TlsContextTrustStoreConfiguration trustStoreConfiguration = tlsContextFactory.getTrustStoreConfiguration();
+
+      if (trustStoreConfiguration != null && trustStoreConfiguration.isInsecure()) {
+        logger.warn(String.format(
+                                  "TLS configuration for requester %s has been set to use an insecure trust store. This means no certificate validations will be performed, rendering connections vulnerable to attacks. Use at own risk.",
+                                  ownerName));
+        // This disables hostname verification
+        builder.setAcceptAnyCertificate(true);
+      }
+    }
+  }
+
+  private void configureProxy(AsyncHttpClientConfig.Builder builder) {
+    if (proxyConfig != null) {
+      doConfigureProxy(builder, proxyConfig);
+    }
+  }
+
+  protected void doConfigureProxy(AsyncHttpClientConfig.Builder builder, ProxyConfig proxyConfig) {
+    builder.setProxyServer(buildProxy(proxyConfig));
+  }
+
+  protected final ProxyServer buildProxy(ProxyConfig proxyConfig) {
+    ProxyServer proxyServer;
+    if (!StringUtils.isEmpty(proxyConfig.getUsername())) {
+      proxyServer =
+          new ProxyServer(proxyConfig.getHost(), proxyConfig.getPort(), proxyConfig.getUsername(), proxyConfig.getPassword());
+      if (proxyConfig instanceof NtlmProxyConfig) {
+        proxyServer.setNtlmDomain(((NtlmProxyConfig) proxyConfig).getNtlmDomain());
+        try {
+          proxyServer.setNtlmHost(getHostName());
+        } catch (UnknownHostException e) {
+          // do nothing, let the default behaviour be used
+        }
+        proxyServer.setScheme(NTLM);
+      }
+    } else {
+      proxyServer = new ProxyServer(proxyConfig.getHost(), proxyConfig.getPort());
+    }
+    return proxyServer;
+  }
+
+  private void configureTransport(AsyncHttpClientConfig.Builder builder) {
+    GrizzlyAsyncHttpProviderConfig providerConfig = new GrizzlyAsyncHttpProviderConfig();
+    CompositeTransportCustomizer compositeTransportCustomizer = new CompositeTransportCustomizer();
+    compositeTransportCustomizer
+        .addTransportCustomizer(new IOStrategyTransportCustomizer(selectorScheduler, workerScheduler));
+    compositeTransportCustomizer.addTransportCustomizer(new LoggerTransportCustomizer());
+
+    if (clientSocketProperties != null) {
+      compositeTransportCustomizer.addTransportCustomizer(new SocketConfigTransportCustomizer(clientSocketProperties));
+    }
+
+    providerConfig.addProperty(GrizzlyAsyncHttpProviderConfig.Property.TRANSPORT_CUSTOMIZER, compositeTransportCustomizer);
+    // Grizzly now decompresses encoded responses, this flag maintains the previous behaviour
+    providerConfig.addProperty(GrizzlyAsyncHttpProviderConfig.Property.DECOMPRESS_RESPONSE, Boolean.FALSE);
+    builder.setAsyncHttpClientProviderConfig(providerConfig);
+  }
+
+  private void configureConnections(AsyncHttpClientConfig.Builder builder) throws InitialisationException {
+    if (maxConnections > 0) {
+      builder.addRequestFilter(new CustomTimeoutThrottleRequestFilter(maxConnections));
+    }
+
+    builder.setMaxConnections(maxConnections);
+    builder.setMaxConnectionsPerHost(maxConnections);
+
+    builder.setAllowPoolingConnections(usePersistentConnections);
+    builder.setAllowPoolingSslConnections(usePersistentConnections);
+
+    builder.setConnectionTTL(MAX_CONNECTION_LIFETIME);
+    builder.setPooledConnectionIdleTimeout(connectionIdleTimeout);
+
+    builder.setIOThreadMultiplier(1);
+  }
+
+  @Override
+  public HttpResponse send(HttpRequest request, int responseTimeout, boolean followRedirects,
+                           HttpRequestAuthentication authentication)
+      throws IOException, TimeoutException {
+
+    Request grizzlyRequest = createGrizzlyRequest(request, responseTimeout, followRedirects, authentication);
+    ListenableFuture<Response> future = asyncHttpClient.executeRequest(grizzlyRequest);
+    try {
+      // No timeout is used to get the value of the future object, as the responseTimeout configured in the request that
+      // is being sent will make the call throw a {@code TimeoutException} if this time is exceeded.
+      Response response = future.get();
+
+      // Under high load, sometimes the get() method returns null. Retrying once fixes the problem (see MULE-8712).
+      if (response == null) {
+        if (logger.isDebugEnabled()) {
+          logger.debug("Null response returned by async client");
+        }
+        response = future.get();
+      }
+      return createMuleResponse(response);
+    } catch (InterruptedException e) {
+      throw new IOException(e);
+    } catch (ExecutionException e) {
+      if (e.getCause() instanceof TimeoutException) {
+        throw (TimeoutException) e.getCause();
+      } else if (e.getCause() instanceof IOException) {
+        throw (IOException) e.getCause();
+      } else {
+        throw new IOException(e);
+      }
+    }
+  }
+
+  @Override
+  public void send(HttpRequest request, int responseTimeout, boolean followRedirects, HttpRequestAuthentication authentication,
+                   final CompletionHandler<HttpResponse, Exception> completionHandler) {
+    try {
+      asyncHttpClient.executeRequest(createGrizzlyRequest(request, responseTimeout, followRedirects, authentication),
+                                     new WorkManagerSourceAsyncCompletionHandler(completionHandler));
+    } catch (Exception e) {
+      completionHandler.onFailure(e);
+    }
+  }
+
+  private class WorkManagerSourceAsyncCompletionHandler extends AsyncCompletionHandler<Response> {
+
+    private CompletionHandler<HttpResponse, Exception> completionHandler;
+
+    WorkManagerSourceAsyncCompletionHandler(CompletionHandler<HttpResponse, Exception> completionHandler) {
+      this.completionHandler = completionHandler;
+    }
+
+    @Override
+    public Response onCompleted(Response response) throws Exception {
+      completionHandler.onCompletion(createMuleResponse(response), this::onThrowable);
+      return null;
+    }
+
+    @Override
+    public void onThrowable(Throwable t) {
+      completionHandler.onFailure((MessagingException) t);
+    }
+
+  }
+
+  private HttpResponse createMuleResponse(Response response) throws IOException {
+    HttpResponseBuilder responseBuilder = HttpResponse.builder();
+    responseBuilder.setStatusCode(response.getStatusCode());
+    responseBuilder.setReasonPhrase(response.getStatusText());
+    responseBuilder.setEntity(new InputStreamHttpEntity(response.getResponseBodyAsStream()));
+
+    if (response.hasResponseHeaders()) {
+      for (String header : response.getHeaders().keySet()) {
+        for (String headerValue : response.getHeaders(header)) {
+          responseBuilder.addHeader(header, headerValue);
+        }
+      }
+    }
+    return responseBuilder.build();
+  }
+
+  private Request createGrizzlyRequest(HttpRequest request, int responseTimeout, boolean followRedirects,
+                                       HttpRequestAuthentication authentication)
+      throws IOException {
+    RequestBuilder reqBuilder = createRequestBuilder(request, builder -> {
+      builder.setMethod(request.getMethod());
+      builder.setFollowRedirects(followRedirects);
+
+      populateHeaders(request, builder);
+
+      for (String queryParamName : request.getQueryParams().keySet()) {
+        for (String queryParamValue : request.getQueryParams().getAll(queryParamName)) {
+          builder.addQueryParam(queryParamName, queryParamValue);
+        }
+      }
+
+      if (authentication != null) {
+        Realm.RealmBuilder realmBuilder = new Realm.RealmBuilder().setPrincipal(authentication.getUsername())
+            .setPassword(authentication.getPassword()).setUsePreemptiveAuth(authentication.isPreemptive());
+
+        if (authentication.getType() == HttpAuthenticationType.BASIC) {
+          realmBuilder.setScheme(Realm.AuthScheme.BASIC);
+        } else if (authentication.getType() == HttpAuthenticationType.DIGEST) {
+          realmBuilder.setScheme(Realm.AuthScheme.DIGEST);
+        } else if (authentication.getType() == HttpAuthenticationType.NTLM) {
+          String domain = authentication.getDomain();
+          if (domain != null) {
+            realmBuilder.setNtlmDomain(domain);
+          }
+          String workstation = authentication.getWorkstation();
+          String ntlmHost = workstation != null ? workstation : getHostName();
+          realmBuilder.setNtlmHost(ntlmHost).setScheme(NTLM);
+        }
+
+        builder.setRealm(realmBuilder.build());
+
+      }
+
+      if (request.getEntity() != null) {
+        if (request.getEntity() instanceof InputStreamHttpEntity) {
+          builder.setBody(new InputStreamBodyGenerator(((InputStreamHttpEntity) request.getEntity()).getInputStream()));
+        } else if (request.getEntity() instanceof ByteArrayHttpEntity) {
+          builder.setBody(((ByteArrayHttpEntity) request.getEntity()).getContent());
+        } else if (request.getEntity() instanceof MultipartHttpEntity) {
+          MultipartHttpEntity multipartHttpEntity = (MultipartHttpEntity) request.getEntity();
+
+          for (HttpPart part : multipartHttpEntity.getParts()) {
+            if (part.getFileName() != null) {
+              builder.addBodyPart(new ByteArrayPart(part.getName(), IOUtils.toByteArray(part.getInputStream()),
+                                                    part.getContentType(), null, part.getFileName()));
+            } else {
+              byte[] content = IOUtils.toByteArray(part.getInputStream());
+              builder.addBodyPart(new ByteArrayPart(part.getName(), content, part.getContentType(), null));
+            }
+          }
+        }
+      }
+
+      // Set the response timeout in the request, this value is read by {@code CustomTimeoutThrottleRequestFilter}
+      // if the maxConnections attribute is configured in the requester.
+      builder.setRequestTimeout(responseTimeout);
+    });
+
+    reqBuilder.setUrl(request.getUri());
+
+    return reqBuilder.build();
+  }
+
+  protected RequestBuilder createRequestBuilder(HttpRequest request, RequestConfigurer requestConfigurer) throws IOException {
+    final RequestBuilder requestBuilder = new RequestBuilder();
+    requestConfigurer.configure(requestBuilder);
+    return requestBuilder;
+  }
+
+  @FunctionalInterface
+  protected interface RequestConfigurer {
+
+    void configure(RequestBuilder reqBuilder) throws IOException;
+  }
+
+  protected void populateHeaders(HttpRequest request, RequestBuilder builder) {
+    for (String headerName : request.getHeaderNames()) {
+      for (String headerValue : request.getHeaderValues(headerName)) {
+        builder.addHeader(headerName, headerValue);
+      }
+    }
+
+    // If persistent connections are disabled, the "Connection: close" header must be explicitly added. AHC will
+    // add "Connection: keep-alive" otherwise. (https://github.com/AsyncHttpClient/async-http-client/issues/885)
+
+    if (!usePersistentConnections) {
+      String connectionHeaderValue = request.getHeaderValueIgnoreCase(CONNECTION);
+      if (connectionHeaderValue != null && !CLOSE.equals(connectionHeaderValue) && logger.isDebugEnabled()) {
+        logger.debug("Persistent connections are disabled in the HTTP requester configuration, but the request already "
+            + "contains a Connection header with value {}. This header will be ignored, and a Connection: close header "
+            + "will be sent instead.", connectionHeaderValue);
+      }
+      builder.setHeader(CONNECTION, CLOSE);
+    }
+  }
+
+  private String getHostName() throws UnknownHostException {
+    return InetAddress.getLocalHost().getHostName();
+  }
+
+  protected ProxyConfig getProxyConfig() {
+    return proxyConfig;
+  }
+
+  @Override
+  public void stop() {
+    asyncHttpClient.close();
+    workerScheduler.stop(5, SECONDS);
+    selectorScheduler.stop(5, SECONDS);
+  }
+}
