--- modules/http/src/main/java/org/mule/runtime/module/http/internal/request/DefaultProxyConfig.java
@@ -1,62 +1,0 @@
-/*
- * Copyright (c) MuleSoft, Inc.  All rights reserved.  http://www.mulesoft.com
- * The software in this package is published under the terms of the CPAL v1.0
- * license, a copy of which has been included with this distribution in the
- * LICENSE.txt file.
- */
-package org.mule.runtime.module.http.internal.request;
-
-import org.mule.service.http.api.client.proxy.ProxyConfig;
-
-/**
- * Basic HTTP Proxy configuration based on host and port, and optionally a username and password for proxy authentication.
- */
-public class DefaultProxyConfig implements ProxyConfig {
-
-  private String name;
-  private String host;
-  private int port = Integer.MAX_VALUE;
-  private String username;
-  private String password;
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public String getHost() {
-    return host;
-  }
-
-  public void setHost(String host) {
-    this.host = host;
-  }
-
-  public int getPort() {
-    return port;
-  }
-
-  public void setPort(int port) {
-    this.port = port;
-  }
-
-  public String getUsername() {
-    return username;
-  }
-
-  public void setUsername(String username) {
-    this.username = username;
-  }
-
-  public String getPassword() {
-    return password;
-  }
-
-  public void setPassword(String password) {
-    this.password = password;
-  }
-
-}
