+++ modules/http/src/test/java/org/mule/runtime/module/http/internal/request/grizzly/LoggerTransportCustomizerTestCase.java
@@ -11,8 +11,9 @@ import static org.mockito.Matchers.isA;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import org.mule.compatibility.module.http.internal.HttpMessageLogger;
+import org.mule.compatibility.module.http.internal.request.grizzly.LoggerTransportCustomizer;
 import org.mule.runtime.api.exception.MuleRuntimeException;
-import org.mule.runtime.module.http.internal.HttpMessageLogger;
 import org.mule.tck.junit4.AbstractMuleTestCase;
 import org.mule.tck.size.SmallTest;
